Estimate Pi
I added the code to run the simulation to estimate pi.

The general idea is that we draw the unit circle and randomly select
a point on the x,y axis. The ratio of points inside of the unit circle
vs the points outside of the unit circle is pi.

NOTE: here is a small performance gain by only using the first quadrant
of the unit circle and multiplying the result by 4.